chore: 메인페이지 코드 분리
