+++ src/components/Home/TopLayer.tsx
@@ -1,0 +1,27 @@
+import * as S from "@/styles/index/index.styles";
+import React from "react";
+
+function TopLayer() {
+  return (
+    <S.ImageWrapper>
+      <S.DesktopImage
+        src="/image/homeDesktop2.jpg"
+        alt="/image/homeDesktop2.jpg"
+        layout="fill"
+        objectFit="cover"
+      />
+      <S.MobileImage
+        src="/image/homeMobile1.jpg"
+        alt="/image/homeMobile1.jpg"
+        layout="fill"
+        objectFit="cover"
+      />
+      <S.TextWrapper>
+        <S.Text>코디로 이해시키는 제품의 가치 </S.Text>
+        <S.Text2>RECORDY SLOW</S.Text2>
+        <S.StyledButton title="view more" width="10rem" />
+      </S.TextWrapper>
+    </S.ImageWrapper>
+  );
+}
+export default TopLayer;
